added data formating utils
